+++ hata/discord/guild/audit_logs/change_converters/channel.py
@@ -1,6 +1,6 @@
 __all__ = ()
 
-from ....channel import ChannelFlag, ForumTag, VideoQualityMode, VoiceRegion
+from ....channel import ChannelFlag, ForumTag, SortOrder, VideoQualityMode, VoiceRegion
 from ....emoji import create_emoji_from_exclusive_data
 from ....permission import PermissionOverwrite
 
@@ -119,12 +119,17 @@ def convert_default_thread_reaction(name, data):
     return AuditLogChange('default_thread_reaction', before, after) 
 
 
+def default_sort_order(name, data):
+    return _convert_preinstanced('default_sort_order', data, SortOrder)
+
+
 CHANNEL_CONVERTERS = {
     'applied_tags': convert_snowflake_array__applied_tag_ids,
     'archived': convert_nothing,
     'auto_archive_duration': convert_int__auto_archive_after,
     'available_tags': convert_forum_tags,
     'bitrate': convert_nothing,
+    'default_sort_order': default_sort_order,
     'default_auto_archive_duration': convert_int__default_thread_auto_archive_after,
     'default_reaction_emoji': convert_default_thread_reaction,
     'default_thread_rate_limit_per_user': convert_int__default_thread_slowmode,